requirements.txt
fastapi==0.75.2
starlette==0.17.1
numpy==1.22.3
python-dotenv==0.20.0
pydantic==1.9.0
uvicorn==0.17.6